+++ CHANGELOG.md
@@ -2,3 +2,5 @@ Changes since 0.9.0
 ===================
 
 * FEATURE: added `cmd/freezer file cat` to write files, in plaintext, straight to stdio.
+* FEATURE: added `cmd/freezer --version` support
+* BREAKING: changed two previous `--version` flags to `file cat --catver` and `sync --syncver`